@@ -57,20 +57,20 @@
 		"number-flow": "workspace:*"
 	},
 	"peerDependencies": {
-		"svelte": "^4 || ^5"
+		"svelte": "^5"
 	},
 	"devDependencies": {
 		"@playwright/test": "^1.45.3",
 		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
+		"@sveltejs/kit": "^2.47.0",
 		"@sveltejs/package": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/vite-plugin-svelte": "^6.0.0",
 		"autoprefixer": "^10.4.20",
 		"publint": "^0.2.0",
-		"svelte": "^4.0.0",
+		"svelte": "^5.0.0",
 		"svelte-check": "^4.0.0",
 		"tailwindcss": "^3.4.9",
 		"typescript": "^5.0.0",
-		"vite": "^5.0.11"
+		"vite": "^6.3.0"
 	}
 }
