@@ -1,5 +1,7 @@
 <script lang="ts">
 	import '../app.css'
+
+	let { children } = $props()
 </script>
 
-<slot></slot>
+{@render children()}
