@@ -1,15 +1,13 @@
 <script lang="ts">
 	import NumberFlow, { NumberFlowGroup, NumberFlowElement, continuous } from '$lib/index.js'
-	import { afterUpdate, tick } from 'svelte'
 
 	const initialValue = 42
 
-	let value = initialValue
-	let el1: NumberFlowElement | undefined
-	let el2: NumberFlowElement | undefined
+	let value = $state(initialValue)
+	let el1: NumberFlowElement | undefined = $state()
+	let el2: NumberFlowElement | undefined = $state()
 
-	afterUpdate(async () => {
-		await tick()
+	$effect(() => {
 		if (value !== initialValue) {
 			;[
 				...(el1?.shadowRoot?.getAnimations() ?? []),
@@ -28,22 +26,20 @@
 		<NumberFlow
 			bind:el={el1}
 			id="flow1"
-			data-testid="flow1"
 			{value}
 			format={{ style: 'currency', currency: 'USD' }}
 			locales="zh-CN"
 			trend={() => -1}
 			prefix=":"
 			suffix="/mo"
-			on:animationsstart={() => console.log('start')}
-			on:animationsfinish={() => console.log('finish')}
+			onanimationsstart={() => console.log('start')}
+			onanimationsfinish={() => console.log('finish')}
 			transformTiming={{ easing: 'linear', duration: 500 }}
 			spinTiming={{ easing: 'linear', duration: 800 }}
 			opacityTiming={{ easing: 'linear', duration: 500 }}
 		/><NumberFlow
 			bind:el={el2}
 			id="flow2"
-			data-testid="flow2"
 			{value}
 			respectMotionPreference={false}
 			plugins={[continuous]}
@@ -54,10 +50,10 @@
 		/>
 	</NumberFlowGroup>
 </div>
-<button on:click={() => (value = 152)}>Change and pause</button>
+<button onclick={() => (value = 152)}>Change and pause</button>
 <br />
 <button
-	on:click={() => {
+	onclick={() => {
 		;[
 			...(el1?.shadowRoot?.getAnimations() ?? []),
 			...(el2?.shadowRoot?.getAnimations() ?? [])
